made about us page.
was more difficult than i thought.or rather, there were many options. in the first section there is a picture to the right of the paragraph. i´ve been told to not use float and instead use flex or grid. i tried all options and ended up with choosing float since to me that makes the most sense anyway.